Move log limit to a shared package
